pyproject.toml
[project]
name = "bossmediagrouptest"
version = "0.1.0"
description = ""
authors = [
    {name = "SkuldFuture",email = "Vakadana@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pygelf (>=0.4.2,<0.5.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.black]
line-length = 88
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

